renamed view "Home" to "Inventory"
