+++ frontend/src/views/Inventory.vue
